requirements.txt
langchain==0.2.17
langchain-openai==0.1.25
langchain-community==0.2.16
chromadb==0.5.5
python-dotenv==1.0.1
